requirements.txt
Pillow
pyserial
